requirements.txt
python-telegram-bot==21.0
python-dotenv==1.0.0
aiosqlite==0.19.0

flask==3.0.0
httpx>=0.27.0

pytest==8.0.0
pytest-asyncio>=0.24.0
pytest-cov==4.1.0
